ch4/ofi: Un-inline MPIDI_OFI_nopack_putget
This is a heavyweight function, so no need to inline.
