+++ src/mpid/ch4/netmod/ofi/Makefile.mk
@@ -11,6 +11,7 @@ mpi_core_sources   += src/mpid/ch4/netmod/ofi/func_table.c \
                       src/mpid/ch4/netmod/ofi/ofi_comm.c \
                       src/mpid/ch4/netmod/ofi/ofi_datatype.c \
                       src/mpid/ch4/netmod/ofi/ofi_op.c \
+                      src/mpid/ch4/netmod/ofi/ofi_rma.c \
                       src/mpid/ch4/netmod/ofi/ofi_spawn.c \
                       src/mpid/ch4/netmod/ofi/ofi_win.c \
                       src/mpid/ch4/netmod/ofi/ofi_events.c \
